% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{clip_swf_to_hex}
\alias{clip_swf_to_hex}
\title{Function 6: Clipping the SWF data to the extent of one selected hexagon}
\usage{
clip_swf_to_hex(swf_sf, hex_grid, hex_id)
}
\arguments{
\item{swf_sf}{The SWF data as sf-polygons; created in function 1: load_swf_data()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   The SWF data gets clipped to the extent of one selected hexagon
}\if{html}{\out{</div>}}
}
\description{
Function 6: Clipping the SWF data to the extent of one selected hexagon
}
\examples{
      clip_swf_to_hex("my_data", "my_grid", "11")
      clip_swf_to_hex(swf_sf = "my_data", hex_grid = "my_grid", hex_id = "15")
}
