% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{count_hedge_obj}
\alias{count_hedge_obj}
\title{Function 11: Counting the hedge/SWF objects in one selected hexagon}
\usage{
count_hedge_obj(swf_clipped, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Integer. Number of distinct hedge/SWF patches
}\if{html}{\out{</div>}}
}
\description{
Function 11: Counting the hedge/SWF objects in one selected hexagon
}
\examples{
         count_hedge_obj("data_clip", "my_grid", "11")
         count_hedge_obj(swf_clipped = "data_clip", hex_grid = "my_grid", hex_id = "15")
}
