% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{create_hex_grid}
\alias{create_hex_grid}
\title{Function 2: Creating a hexagon grid}
\usage{
create_hex_grid(swf_sf, diameter)
}
\arguments{
\item{swf_sf}{An \code{sf} object of SWF polygons; created in function 1: load_swf_data()}

\item{diameter}{Numeric. Diameter of the hexagons in map units (e.g., meters)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   An `sf` object containing the hexagon grid with assigned IDs
}\if{html}{\out{</div>}}
}
\description{
Function 2: Creating a hexagon grid
}
\examples{
     create_hex_grid("my_data", "500")
     create_hex_grid(swf_sf = "my_data", diameter = "500")
}
