% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedge_area}
\alias{hedge_area}
\title{Function 10: Calculating the hedge area in one selected hexagon}
\usage{
hedge_area(swf_clipped, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The overall area of hedges/SWF in the selected hexagon as well as the percentage of it to the whole hexagon area gets printed
}\if{html}{\out{</div>}}
}
\description{
Function 10: Calculating the hedge area in one selected hexagon
}
\examples{
         hedge_area("data_clip", "my_grid", "11")
         hedge_area(swf_clipped = "data_clip", hex_grid = "my_grid", hex_id = "15")
}
