% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedge_points_percentage}
\alias{hedge_points_percentage}
\title{Function 12: Calculating the percentage of the random points, that lay within hedge/SWF objects}
\usage{
hedge_points_percentage(swf_clipped, random_points, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{random_points}{The random points; created in function 8: random_points()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        The percentage of points within hedge/SWF objects gets printed
}\if{html}{\out{</div>}}
}
\description{
Function 12: Calculating the percentage of the random points, that lay within hedge/SWF objects
}
\examples{
           hedge_point_percentage("data_clip", "my_points", "my_grid", "11")
           hedge_point_percentage(swf_clipped = "data_clip", random_points = "my_points", hex_grid = "my_grid", hex_id = "15")
}
