% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedges_nn}
\alias{hedges_nn}
\title{Function 14: Calculating the mean, min & max distance from all hedge/SWF objects to their x nearest neighbour}
\usage{
hedges_nn(swf_clipped, hex_grid, hex_id, nn)
}
\arguments{
\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}

\item{nn}{Numeric. Number of nearest neighbors to consider}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list with mean, min, and max distances to the x nearest neighbor
}\if{html}{\out{</div>}}
}
\description{
Function 14: Calculating the mean, min & max distance from all hedge/SWF objects to their x nearest neighbour
}
\examples{
         hedges_nn("data_clip", "my_grid", "11", "2")
         hedges_nn(swf_clipped = "data_clip", hex_grid = "my_grid", hex_id = "15", nn = "7")
}
