% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{load_swf_data}
\alias{load_swf_data}
\title{Function 1: Loading in the SWF data}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{  COPERNICUS LAND MONITORING SERVICE (2018): High Resolution Layer Small Woody Features.URL:https://land.copernicus.eu/en/products/high-resolution-layer-small-woody-features.
}\if{html}{\out{</div>}}
}
\usage{
load_swf_data(path, CRS)
}
\arguments{
\item{path}{Character. Path to the SWF raster file (.tif)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ An `sf` object with hedge/SWF polygons
}\if{html}{\out{</div>}}
}
\description{
Data from the SWF Layer from Copernicus
}
\examples{
    load_swf_data(system.file("extdata", "HRL_Small_Woody_Features_2018_005m.tif", package = "HedgeConnectivityPackage"), 3035)

    load_swf_data(path = system.file("extdata", "HRL_Small_Woody_Features_2018_005m.tif", package = "HedgeConnectivityPackage"), CRS = 3035)

}
