% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_random_points}
\alias{plot_random_points}
\title{Function 9: Plotting random points over clipped SWF data}
\usage{
plot_random_points(hex_grid, swf_clipped, hex_id, random_points)
}
\arguments{
\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}

\item{random_points}{The random points; created in function 8: random_points()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        The random points get plotted over the selected SWF data clip
}\if{html}{\out{</div>}}
}
\description{
Function 9: Plotting random points over clipped SWF data
}
\examples{
           plot_random_points("my_grid", "data_clip", "11", "my_points")
           plot_random_points(hex_grid = "my_grid", swf_clipped = "data_clip", hex_id = "11", random_points = "my_points")
}
