% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_swf_hex}
\alias{plot_swf_hex}
\title{Function 7: Plotting the clipped SWF data with the corresponding hexagon}
\usage{
plot_swf_hex(hex_grid, swf_clipped, hex_id)
}
\arguments{
\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A plot showing the SWF polygons inside a hexagon
}\if{html}{\out{</div>}}
}
\description{
Function 7: Plotting the clipped SWF data with the corresponding hexagon
}
\examples{
         plot_swf_hex("my_grid", "data_clip", "11")
         plot_swf_hex(hex_grid = "my_grid", swf_clip = "data_clip", hex_id = "11")
}
