% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{random_points}
\alias{random_points}
\title{Function 8: Creating random points}
\usage{
random_points(hex_grid, hex_id, n_points)
}
\arguments{
\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}

\item{n_points}{Numeric. Number of random points}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   An `sf` object of random points
}\if{html}{\out{</div>}}
}
\description{
Function 8: Creating random points
}
\examples{
      random_points("my_grid", "11", "20")
      random_points(hex_grid = "my_grid", hex_id = "15", n_points = "100")
}
