% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{count_hedge_obj}
\alias{count_hedge_obj}
\title{Function 11: Counting the hedge/SWF objects in one selected hexagon}
\usage{
count_hedge_obj(swf_clipped, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The number of hedge/SWF objects/patches gets printed
}\if{html}{\out{</div>}}
}
\description{
Function 11: Counting the hedge/SWF objects in one selected hexagon
}
\examples{
         count_hedge_obj("data_clip", "my_grid", "11")
         count_hedge_obj(swf_clipped = "data_clip", hex_grid = "my_grid", hex_id = "15")
}
