% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{create_hex_grid}
\alias{create_hex_grid}
\title{Function 2: Creating a hexagon grid}
\usage{
create_hex_grid(swf_sf, diameter)
}
\arguments{
\item{swf_sf}{The SWF data as sf-polygons; created in function 1: load_swf_data()}

\item{diameter}{Diameter of the polygons in the respective unit; freely selectable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A hexagon-grid with the chosen diameter is created
}\if{html}{\out{</div>}}
}
\description{
Function 2: Creating a hexagon grid
}
\examples{
     create_hex_grid("my_data", "500")
     create_hex_grid(swf_sf = "my_data", diameter = "500")
}
