% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{distance_to_nearest_hedge}
\alias{distance_to_nearest_hedge}
\title{Function 13: Calculating the mean, min, max distance from all points outside of hedge/SWF objects to the closest hedge/SWF object}
\usage{
distance_to_nearest_hedge(swf_clipped, random_points, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{random_points}{The random points; created in function 8: random_points()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        The mean, min & max distance from all points outside of hedge/SWF objects in the selected hexagon to the closest hedge/SWF object get printed in a list
}\if{html}{\out{</div>}}
}
\description{
Function 13: Calculating the mean, min, max distance from all points outside of hedge/SWF objects to the closest hedge/SWF object
}
\examples{
           distance_to_nearest_hedge("data_clip", "my_points", "my_grid", "11")
           distance_to_nearest_hedge(swf_clipped = "data_clip", random_points = "my_points", hex_grid = "my_grid", hex_id = "15")
}
