% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{distance_to_nearest_hedge}
\alias{distance_to_nearest_hedge}
\title{Function 13: Calculating the mean, min, max distance from all points outside of hedge/SWF objects to the closest hedge/SWF object}
\usage{
distance_to_nearest_hedge(swf_clipped, random_points, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{random_points}{An \code{sf} object of random points; created in function 8: random_points()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list with mean, min and max distances for points outside of hedges
}\if{html}{\out{</div>}}
}
\description{
Function 13: Calculating the mean, min, max distance from all points outside of hedge/SWF objects to the closest hedge/SWF object
}
\examples{
           my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                     my_grid <- create_hex_grid(my_data, 500)
                     data_clip <- swf_grid(my_data, my_grid)
                     my_points <- random_points(my_grid, 11, 20)

                     distance_to_nearest_hedge(data_clip, my_points, my_grid, 11)

           my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                     my_grid <- create_hex_grid(my_data, 500)
                     data_clip <- swf_grid(my_data, my_grid)
                     my_points <- random_points(my_grid, 15, 25)

                     distance_to_nearest_hedge(swf_clipped = data_clip, random_points = my_points, hex_grid = my_grid, hex_id = 15)
}
