% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedge_area}
\alias{hedge_area}
\title{Function 10: Calculating the hedge area in one selected hexagon}
\usage{
hedge_area(swf_clipped, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list with total hedge area and area ratio
}\if{html}{\out{</div>}}
}
\description{
Function 10: Calculating the hedge area in one selected hexagon
}
\examples{
         my_data <- load_swf_data(system.file("extdata", "HRL_Small_Woody_Features_2018_005m.tif", package = "HedgeConnectivityPackage"), 3035)
                   my_grid <- create_hex_grid(my_data, 500)
                   data_clip <- swf_grid(my_data, my_grid)

                   hedge_area(data_clip, my_grid, 11)

         my_data <- load_swf_data(system.file("extdata", "HRL_Small_Woody_Features_2018_005m.tif", package = "HedgeConnectivityPackage"), 3035)
                   my_grid <- create_hex_grid(my_data, 500)
                   data_clip <- swf_grid(my_data, my_grid)

                   hedge_area(swf_clipped = data_clip, hex_grid = my_grid, hex_id = 15)
}
