% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedge_points_percentage}
\alias{hedge_points_percentage}
\title{Function 12: Calculating the percentage of the random points, that lay within hedge/SWF objects}
\usage{
hedge_points_percentage(swf_clipped, random_points, hex_grid, hex_id)
}
\arguments{
\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{random_points}{An \code{sf} object of random points; created in function 8: random_points()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Numeric. Percentage of random points inside hedge/SWF polygons
}\if{html}{\out{</div>}}
}
\description{
Function 12: Calculating the percentage of the random points, that lay within hedge/SWF objects
}
\examples{
           my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                     my_grid <- create_hex_grid(my_data, 500)
                     data_clip <- swf_grid(my_data, my_grid)
                     my_points <- random_points(my_grid, 11, 20)

                     hedge_points_percentage(data_clip, my_points, my_grid, 11)

           my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                     my_grid <- create_hex_grid(my_data, 500)
                     data_clip <- swf_grid(my_data, my_grid)
                     my_points <- random_points(my_grid, 15, 25)

                     hedge_points_percentage(swf_clipped = data_clip, random_points = my_points, hex_grid = my_grid, hex_id = 15)
}
