% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{hedges_nn}
\alias{hedges_nn}
\title{Function 14: Calculating the mean, min & max distance from all hedge/SWF objects to their x nearest neighbour}
\usage{
hedges_nn(swf_clipped, hex_grid, hex_id, nn)
}
\arguments{
\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}

\item{nn}{The nearest neighbour, to which the distances should be calculated}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The mean, min & max distance from all hedge/SWF objects in the selected hexagon to the x nearest neighbour
}\if{html}{\out{</div>}}
}
\description{
Function 14: Calculating the mean, min & max distance from all hedge/SWF objects to their x nearest neighbour
}
\examples{
         hedges_nn("data_clip", "my_grid", "11", "2")
         hedges_nn(swf_clipped = "data_clip", hex_grid = "my_grid", hex_id = "15", nn = "7")
}
