% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{load_swf_data}
\alias{load_swf_data}
\title{Function 1: Loading in the SWF data}
\usage{
load_swf_data(path)
}
\arguments{
\item{path}{Character. Path to the SWF raster file (.tif)}
}
\value{
An \code{sf} object with hedge/SWF polygons
}
\description{
Function 1: Loading in the SWF data
}
\examples{
  load_swf_data("your/individual/path/file.tif")
  load_swf_data(path = "your/individual/path/file.tif")
}
