% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_hex_ids}
\alias{plot_hex_ids}
\title{Function 3: Plotting the hexagon grid with IDs}
\usage{
plot_hex_ids(hex_grid)
}
\arguments{
\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   An plot showing the hexagons labeled by ID
}\if{html}{\out{</div>}}
}
\description{
Function 3: Plotting the hexagon grid with IDs
}
\examples{
      my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif")
                my_grid <- create_hex_grid(my_data, 500)

                plot_hex_ids(my_grid)

      my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif")
                my_grid <- create_hex_grid(my_data, 500)

                plot_hex_ids(hex_grid = my_grid)
}
