% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_hex_ids}
\alias{plot_hex_ids}
\title{Function 3: Plotting the hexagon grid with IDs}
\usage{
plot_hex_ids(hex_grid)
}
\arguments{
\item{hex_grid}{The created hexagon grid; created in function 2: create_hex_grid()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A plot of the hexagon grid with the according IDs is plotted
}\if{html}{\out{</div>}}
}
\description{
Function 3: Plotting the hexagon grid with IDs
}
\examples{
      plot_hex_ids("my_grid")
      plot_hex_ids(hex_grid = "my_grid")
}
