% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_random_points}
\alias{plot_random_points}
\title{Function 9: Plotting random points over clipped SWF data}
\usage{
plot_random_points(hex_grid, swf_clipped, hex_id, random_points)
}
\arguments{
\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}

\item{random_points}{An \code{sf} object of random points; created in function 8: random_points()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A plot showing random points over clipped SWF polygons
}\if{html}{\out{</div>}}
}
\description{
Function 9: Plotting random points over clipped SWF data
}
\examples{
           plot_random_points("my_grid", "data_clip", "11", "my_points")
           plot_random_points(hex_grid = "my_grid", swf_clipped = "data_clip", hex_id = "11", random_points = "my_points")
}
