% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_swf_grid}
\alias{plot_swf_grid}
\title{Function 5: Plotting the whole clipped SWF data with the hexagon grid}
\usage{
plot_swf_grid(hex_grid, swf_clipped)
}
\arguments{
\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The clipped SWF data together with the hexagon grid gets plotted
}\if{html}{\out{</div>}}
}
\description{
Function 5: Plotting the whole clipped SWF data with the hexagon grid
}
\examples{
         plot_swf_grid("my_grid", "data_clip")
         plot_swf_grid(hex_grid = "my_grid", swf_clipped = "data_clip")
}
