% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_swf_hex}
\alias{plot_swf_hex}
\title{Function 7: Plotting the clipped SWF data with the corresponding hexagon}
\usage{
plot_swf_hex(hex_grid, swf_clipped, hex_id)
}
\arguments{
\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{swf_clipped}{The clipped SWF data; created in function 4: swf_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      The SWF data clipped to the extent of one selected hexagon gets plotted together with the hexagon boarders
}\if{html}{\out{</div>}}
}
\description{
Function 7: Plotting the clipped SWF data with the corresponding hexagon
}
\examples{
         plot_swf_hex("my_grid", "data_clip", "11")
         plot_swf_hex(hex_grid = "my_grid", swf_clip = "data_clip", hex_id = "11")
}
