% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{plot_swf_hex}
\alias{plot_swf_hex}
\title{Function 7: Plotting the clipped SWF data with the corresponding hexagon}
\usage{
plot_swf_hex(hex_grid, swf_clipped, hex_id)
}
\arguments{
\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}

\item{swf_clipped}{An \code{sf} object of clipped SWF polygons; created in function 4: swf_grid()}

\item{hex_id}{Numeric. ID of the selected hexagon}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A plot showing the SWF polygons inside a hexagon
}\if{html}{\out{</div>}}
}
\description{
Function 7: Plotting the clipped SWF data with the corresponding hexagon
}
\examples{
         my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                   my_grid <- create_hex_grid(my_data, 500)
                   data_clip <- swf_grid(my_data, my_grid)

                   plot_swf_hex(my_grid, data_clip, 11)

         my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif", 3035)
                   my_grid <- create_hex_grid(my_data, 500)
                   data_clip <- swf_grid(my_data, my_grid)

                   plot_swf_hex(hex_grid = my_grid, swf_clipped = data_clip, hex_id = 15)
}
