% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{random_points}
\alias{random_points}
\title{Function 8: Creating random points}
\usage{
random_points(hex_grid, hex_id, n_points)
}
\arguments{
\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}

\item{hex_id}{The ID of the hexagon that should be selected}

\item{n_points}{The number of random points}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   Random points over the extent of the selected hexagon are created
}\if{html}{\out{</div>}}
}
\description{
Function 8: Creating random points
}
\examples{
      random_points("my_grid", "11", "20")
      random_points(hex_grid = "my_grid", hex_id = "15", n_points = "100")
}
