% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{swf_grid}
\alias{swf_grid}
\title{Function 4: Clipping the whole SWF data with the hexagon grid}
\usage{
swf_grid(swf_sf, hex_grid)
}
\arguments{
\item{swf_sf}{An \code{sf} object of SWF polygons; created in function 1: load_swf_data()}

\item{hex_grid}{An \code{sf} object of hexagons with assigned IDs; created in function 2: create_hex_grid()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   An `sf` object of clipped SWF data
}\if{html}{\out{</div>}}
}
\description{
Function 4: Clipping the whole SWF data with the hexagon grid
}
\examples{
      my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif")
                my_grid <- create_hex_grid(my_data, 500)

                swf_grid(my_data, my_grid)

      my_data <- load_swf_data("inst/extdata/HRL_Small_Woody_Features_2018_005m.tif")
                my_grid <- create_hex_grid(my_data, 500)

                swf_grid(swf_sf = my_data, hex_grid = my_grid)
}
