% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_test.R
\name{swf_grid}
\alias{swf_grid}
\title{Function 4: Clipping the whole SWF data with the hexagon grid}
\usage{
swf_grid(swf_sf, hex_grid)
}
\arguments{
\item{swf_sf}{The SWF data as sf-polygons; created in function 1: load_swf_data()}

\item{hex_grid}{The hexagon grid; created in function 2: create_hex_grid()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   The SWF data gets clipped with the hexagon grid; no visible output
}\if{html}{\out{</div>}}
}
\description{
Function 4: Clipping the whole SWF data with the hexagon grid
}
\examples{
      swf_grid("my_data", "my_grid")
      swf_grid(swf_sf = "my_data", hex_grid = "my_grid")
}
